+++ sdk/appcenter-storage/src/main/java/com/microsoft/appcenter/storage/client/TokenExchange.java
@@ -7,9 +7,8 @@ package com.microsoft.appcenter.storage.client;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import com.microsoft.appcenter.AppCenter;
+import com.microsoft.appcenter.http.AbstractAppCallTemplate;
 import com.microsoft.appcenter.http.HttpClient;
-import com.microsoft.appcenter.http.HttpUtils;
 import com.microsoft.appcenter.http.ServiceCall;
 import com.microsoft.appcenter.http.ServiceCallback;
 import com.microsoft.appcenter.storage.Constants;
@@ -21,11 +20,9 @@ import com.microsoft.appcenter.storage.models.TokensResponse;
 import com.microsoft.appcenter.utils.AppCenterLog;
 import com.microsoft.appcenter.utils.context.AuthTokenContext;
 
-import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 
-import static android.util.Log.VERBOSE;
 import static com.microsoft.appcenter.Constants.APP_SECRET;
 import static com.microsoft.appcenter.Constants.AUTHORIZATION_HEADER;
 import static com.microsoft.appcenter.Constants.AUTH_TOKEN_FORMAT;
@@ -84,32 +81,11 @@ public class TokenExchange {
                         }
                     }
                 },
-                new HttpClient.CallTemplate() {
+                new AbstractAppCallTemplate() {
 
                     @Override
                     public String buildRequestBody() {
                         return buildAppCenterGetDbTokenBodyPayload(partition);
-                    }
-
-                    @Override
-                    public void onBeforeCalling(URL url, Map<String, String> headers) {
-                        if (AppCenterLog.getLogLevel() <= VERBOSE) {
-
-                            /* Log url. */
-                            AppCenterLog.verbose(AppCenter.LOG_TAG, "Calling " + url + "...");
-
-                            /* Log headers. */
-                            Map<String, String> logHeaders = new HashMap<>(headers);
-                            String appSecret = logHeaders.get(APP_SECRET);
-                            if (appSecret != null) {
-                                logHeaders.put(APP_SECRET, HttpUtils.hideSecret(appSecret));
-                            }
-                            String authToken = logHeaders.get(AUTHORIZATION_HEADER);
-                            if (authToken != null) {
-                                logHeaders.put(AUTHORIZATION_HEADER, HttpUtils.hideAuthToken(authToken));
-                            }
-                            AppCenterLog.verbose(AppCenter.LOG_TAG, "Headers: " + logHeaders);
-                        }
                     }
                 },
                 serviceCallback);
