+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/ingestion/AppCenterIngestion.java
@@ -10,24 +10,22 @@ import android.support.annotation.NonNull;
 import android.support.annotation.VisibleForTesting;
 
 import com.microsoft.appcenter.Constants;
+import com.microsoft.appcenter.http.AbstractAppCallTemplate;
 import com.microsoft.appcenter.http.HttpClient;
-import com.microsoft.appcenter.http.HttpUtils;
 import com.microsoft.appcenter.http.ServiceCall;
 import com.microsoft.appcenter.http.ServiceCallback;
 import com.microsoft.appcenter.ingestion.models.LogContainer;
 import com.microsoft.appcenter.ingestion.models.json.LogSerializer;
-import com.microsoft.appcenter.utils.AppCenterLog;
 
 import org.json.JSONException;
 
 import java.io.IOException;
-import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
-import static android.util.Log.VERBOSE;
-import static com.microsoft.appcenter.AppCenter.LOG_TAG;
+import static com.microsoft.appcenter.Constants.APP_SECRET;
+import static com.microsoft.appcenter.Constants.AUTHORIZATION_HEADER;
 import static com.microsoft.appcenter.http.DefaultHttpClient.METHOD_POST;
 import static com.microsoft.appcenter.http.HttpUtils.createHttpClient;
 
@@ -108,10 +106,10 @@ public class AppCenterIngestion implements Ingestion {
     public ServiceCall sendAsync(String appSecret, UUID installId, LogContainer logContainer, final ServiceCallback serviceCallback) throws IllegalArgumentException {
         Map<String, String> headers = new HashMap<>();
         headers.put(INSTALL_ID, installId.toString());
-        headers.put(Constants.APP_SECRET, appSecret);
+        headers.put(APP_SECRET, appSecret);
         String authToken = getAuthToken();
         if (authToken != null) {
-            headers.put(Constants.AUTHORIZATION_HEADER, String.format(Constants.AUTH_TOKEN_FORMAT, authToken));
+            headers.put(AUTHORIZATION_HEADER, String.format(Constants.AUTH_TOKEN_FORMAT, authToken));
         }
         HttpClient.CallTemplate callTemplate = new IngestionCallTemplate(mLogSerializer, logContainer);
         return mHttpClient.callAsync(mLogUrl + API_PATH, METHOD_POST, headers, callTemplate, serviceCallback);
@@ -130,7 +128,7 @@ public class AppCenterIngestion implements Ingestion {
     /**
      * Inner class is used to be able to mock System.currentTimeMillis, does not work if using anonymous inner class...
      */
-    private static class IngestionCallTemplate implements HttpClient.CallTemplate {
+    private static class IngestionCallTemplate extends AbstractAppCallTemplate {
 
         private final LogSerializer mLogSerializer;
 
@@ -146,27 +144,6 @@ public class AppCenterIngestion implements Ingestion {
 
             /* Serialize payload. */
             return mLogSerializer.serializeContainer(mLogContainer);
-        }
-
-        @Override
-        public void onBeforeCalling(URL url, Map<String, String> headers) {
-            if (AppCenterLog.getLogLevel() <= VERBOSE) {
-
-                /* Log url. */
-                AppCenterLog.verbose(LOG_TAG, "Calling " + url + "...");
-
-                /* Log headers. */
-                Map<String, String> logHeaders = new HashMap<>(headers);
-                String appSecret = logHeaders.get(Constants.APP_SECRET);
-                if (appSecret != null) {
-                    logHeaders.put(Constants.APP_SECRET, HttpUtils.hideSecret(appSecret));
-                }
-                String authToken = logHeaders.get(Constants.AUTHORIZATION_HEADER);
-                if (authToken != null) {
-                    logHeaders.put(Constants.AUTHORIZATION_HEADER, HttpUtils.hideAuthToken(authToken));
-                }
-                AppCenterLog.verbose(LOG_TAG, "Headers: " + logHeaders);
-            }
         }
     }
 }
